+++ src/Symfony/Component/Serializer/Tests/Mapping/Loader/XmlFileLoaderTest.php
@@ -82,6 +82,10 @@ class XmlFileLoaderTest extends TestCase
         $attributesMetadata = $classMetadata->getAttributesMetadata();
         $this->assertEquals('baz', $attributesMetadata['foo']->getSerializedName());
         $this->assertEquals('qux', $attributesMetadata['bar']->getSerializedName());
+
+        $this->assertEquals('nameDefault', $attributesMetadata['quux']->getSerializedName());
+        $this->assertEquals('nameOne', $attributesMetadata['quux']->getSerializedName(['groupB']));
+        $this->assertEquals('nameTwo', $attributesMetadata['quux']->getSerializedName(['groupC']));
     }
 
     public function testSerializedPath()