+++ src/Symfony/Component/Serializer/CHANGELOG.md
@@ -12,6 +12,8 @@ CHANGELOG
  * Change the signature of `ClassMetadataInterface::setClassDiscriminatorMapping()` to `setClassDiscriminatorMapping(?ClassDiscriminatorMapping)`
  * Add option YamlEncoder::YAML_INDENTATION to YamlEncoder constructor options to configure additional indentation for each level of nesting. This allows configuring indentation in the service configuration.
  * Add `SerializedPath` annotation to flatten nested attributes
+ * Add serialized name group support
+ * Add argument `$groups` to `AttributeMetadata::setSerializedName()` and `AttributeMetadata::getSerializedName()`
 
 6.1
 ---